requirements.txt
pyrtlsdr
numpy
scipy
pytest